Fix cadvisor flag registration for cross build
